agent/pyproject.toml
[tool.poetry]
name = "advisor"
version = "0.1.0"
description = ""
authors = ["Tyler Slaton <tyler@copilotkit.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.34"
langchain-core = "^0.3.8"
langchain-openai = "^0.2.1"
langchain-anthropic = "^0.2.1"
langchain = "^0.3.1"
openai = "^1.51.0"
langchain-community = "^0.3.1"
copilotkit = "0.1.27"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
tavily-python = "^0.5.0"
html2text = "^2024.2.26"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = "advisor.server:main"
